Fix error with multiple GPUs
When using multiple GPUs in the same machine, for example GPU 0 and GPU 1, if the model is in GPU 1, the changed line will move the tensors to GPU 0, generating errors because model and tensors are not in the same device.

This is because `device.type` returns just `cuda` and not `cuda:0` or `cuda:1`. As the `device` variable is only used for moving tensors to the appropriated device, using `val.device` do the same job and correctly move the tensors to the desired device.